Move Hook based createAnimatedComponent to a new file and behind a flag
